+++ benches/aco_bench.rs
@@ -1,34 +1,30 @@
+use aco::tsp;
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use ecrs::aco;
-use ecrs::aco::ants_behaviour::AntSystemAB;
-use ecrs::aco::colony::LegacyColony;
-use ecrs::aco::goodness::CanonicalGoodness;
-use ecrs::aco::grader::PathLengthInverse;
-use ecrs::aco::pheromone::AntSystemPU;
-use ecrs::aco::{util, FMatrix};
+use ecrs::aco::FMatrix;
 use itertools::Itertools;
 use std::time::Duration;
 
 pub fn bench_aco_small(c: &mut Criterion) {
     let dist = calc_dist(&CITIES_5);
-    let heuristic = util::create_heuristic_from_weights(&dist);
+    let heuristic = tsp::util::create_heuristic_from_weights(&dist);
     let start_pheromone = FMatrix::repeat(heuristic.nrows(), heuristic.ncols(), 1.0);
 
     c.bench_function("aco small", |b| {
         b.iter(|| {
             let ants = (0..5)
-                .map(|_| aco::ant::CanonicalAnt::new(CITIES_5.len()))
+                .map(|_| tsp::ant::CanonicalAnt::new(CITIES_5.len()))
                 .collect_vec();
-            let colony = LegacyColony::new(
-                AntSystemAB,
-                CanonicalGoodness::new(2.0, 2.0, heuristic.clone()),
+            let colony = tsp::TspColony::new(
+                tsp::ants_behaviour::AntSystemAB,
+                tsp::goodness::CanonicalGoodness::new(2.0, 2.0, heuristic.clone()),
                 ants,
             );
 
-            aco::Builder::new(5)
+            aco::Builder::new()
                 .set_colony(colony)
-                .set_grader(PathLengthInverse::new(black_box(dist.clone())))
-                .set_pheromone_update(AntSystemPU::new(0.1))
+                .set_grader(tsp::PathLengthInverse::new(black_box(dist.clone())))
+                .set_pheromone_update(tsp::pheromone::AntSystemPU::new(0.1))
                 .set_probe(EmptyProbe)
                 .with_iteration_termination(black_box(20))
                 .set_start_pheromone(start_pheromone.clone())
@@ -40,24 +36,24 @@ pub fn bench_aco_small(c: &mut Criterion) {
 
 pub fn bench_aco_medium(c: &mut Criterion) {
     let dist = calc_dist(&CITIES_15);
-    let heuristic = util::create_heuristic_from_weights(&dist);
+    let heuristic = tsp::util::create_heuristic_from_weights(&dist);
     let start_pheromone = FMatrix::repeat(heuristic.nrows(), heuristic.ncols(), 1.0);
 
     c.bench_function("aco medium", |b| {
         b.iter(|| {
             let ants = (0..5)
-                .map(|_| aco::ant::CanonicalAnt::new(CITIES_15.len()))
+                .map(|_| tsp::ant::CanonicalAnt::new(CITIES_15.len()))
                 .collect_vec();
-            let colony = LegacyColony::new(
-                AntSystemAB,
-                CanonicalGoodness::new(2.0, 2.0, heuristic.clone()),
+            let colony = tsp::TspColony::new(
+                tsp::ants_behaviour::AntSystemAB,
+                tsp::goodness::CanonicalGoodness::new(2.0, 2.0, heuristic.clone()),
                 ants,
             );
 
-            aco::Builder::new(15)
-                .set_grader(PathLengthInverse::new(black_box(dist.clone())))
+            aco::Builder::new()
+                .set_grader(tsp::PathLengthInverse::new(black_box(dist.clone())))
                 .set_colony(colony)
-                .set_pheromone_update(AntSystemPU::new(0.1))
+                .set_pheromone_update(tsp::pheromone::AntSystemPU::new(0.1))
                 .set_probe(EmptyProbe)
                 .with_iteration_termination(black_box(20))
                 .set_start_pheromone(start_pheromone.clone())
